// src/backoffice/rescue/dto/create-rescue.dto.ts
export class CreateRescueDto {
  rescue: {
    name: string;
    artists: string;
    cover_url: string;
    uid: string;
    playlist: string;
    limited: boolean;
    info_playlist: string;
    limit_streams: number;
    uri: string;
    score: number;
    status: boolean;
    isrc: string;
    priority: number;
  };
}